requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pydantic-settings
# Optional: Keep these if you want to maintain Telegram bot functionality
# python-telegram-bot
# google-generativeai
